rm old client
